requirements.txt
# AI-Powered Fake News Detection System
# Requirements File

# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# NLP and Text Processing
transformers==4.36.0
torch==2.1.2
torchvision==0.16.2
nltk==3.8.1
spacy==3.7.4

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
scikit-image==0.22.0

# Video Processing
moviepy==1.0.3

# Multilingual Support
sentence-transformers==2.2.2
googletrans==4.0.0-rc1

# Data and Utilities
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0

# Frontend
streamlit==1.30.0

# API Documentation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
